+++ GraphClassification/GraphClassifi/util.py
@@ -1,75 +1,132 @@
-import GCN as md
+# https://github.com/fahim-sikder/Node-Classification-GCN/blob/master/pytorch-GCN.ipynb
+
 import torch
-import util as ut
+import torch.nn as nn
+import torch.nn.functional as F
+import torch.optim as optim
+import numpy as np
+import numpy as np
+import torch
+import math
+from torch.nn.parameter import Parameter
 from torch.nn.modules.module import Module
 import torch.nn as nn
-#import pandas as pd
+import torch.nn.functional as F
+import time
 import torch.optim as optim
+# import pandas as pd
+from scipy.sparse import csr_matrix
+from scipy.sparse import coo_matrix
+from scipy.sparse import diags
+from scipy.sparse import eye
+from pathlib import Path
+from functools import partial
+import sys
 
+'''
+GNN Node Classification
+dataset : Visual Genome.Scence graph 
+
+Adj : id x id (같은 cluster 값을 갖는 경우 1로 체크함), 1000x1000
+Feature : id x freOBJ(해당 img에 FreOBJ가 있는 경우 1, 없는 경우 0)
+freObj : 대상이 되는 이미지 1000개의 Scene graph에서 가장 언급량이 많은 Obj 100개
+Label : 각 id 당 cluster 번호
+cluster는 bert-base-nli-mean-tokens를 이용해 15개의 클러스터로 분류했음
+-> 1000개의 region_graph의 phrase 값을 embedding 함
+-> 이 부분도 relationship이 잘 나타나지 않은 것 같아 아쉬움
+relationship을 더 잘 활용할 수 있는걸 하고싶음..
+
+'''
 
 # gpu 사용
 device = torch.device('cuda') if torch.cuda.is_available() else torch.device('cpu')
 
-features,adj, labels = ut.loadData() #데이터 불러옴
-idx_train, idx_val, idx_test = ut.splitDataset(34,200,300,len(features)) #train/val/test로 나눌 imageId를 가진 배열 return
-ut.toDevice(adj, features, labels, idx_train, idx_val, idx_test) # gpu면 gpu로 cpu면 cpu로~
+
+def normalize(mx):
+    rowsum = np.array(mx.sum(1))
+    r_inv = (rowsum ** -1).flatten()
+    r_inv[np.isinf(r_inv)] = 0.
+    r_mat_inv = diags(r_inv)
+    mx = r_mat_inv.dot(mx)
+    return mx
 
 
-n_labels = labels.max().item() + 1  # 15
-n_features = features.shape[1]  # 100
-# seed 고정
-torch.manual_seed(34)
+def loadData():
+    features = csr_matrix(np.load('./data/idFreFeature.npy'), dtype=np.float32)  # csr_matrix : (1000,100)
+    adj = torch.FloatTensor(np.load('./data/idAdj.npy'))  # tensor(1000,1000)
+    features = csr_matrix(features)
 
-# model
-#GCN(  (gc1): GraphConvolution (100 -> 20)   (gc2): GraphConvolution (20 -> 15) )
-model = md.GCN(nfeat=n_features,
-            nhid=20,  # hidden = 16
-            nclass=n_labels,
-            dropout=0.5)  # dropout = 0.5
-optimizer = optim.Adam(model.parameters(),
-                       lr=0.001, weight_decay=5e-4)
+    features = normalize(features)
 
-epochs = 1000
-print_steps = 100
-train_loss, train_acc = [], []
-val_loss, val_acc = [], []
+    testFile = open('./data/cluster.txt', 'r')  # 'r' read의 약자, 'rb' read binary 약자 (그림같은 이미지 파일 읽을때)
+    readFile = testFile.readline()
+    label = (readFile[1:].replace("'", '').replace(' ', '').split(','))
+    labels = []
+    for i in range(1000):
+        labels.append(int(label[i]))
+
+    features = torch.FloatTensor(
+        features.todense())  # <class 'torch.Tensor'> (1000,100), sparseMatrix -> Tensor Matrix, torch.float32
+    labels = torch.LongTensor(labels)  # <class 'torch.Tensor'> tensor(1000, ) int64
+
+    # floatTensor, floatTensor, LongTensor로 반환
+
+    return features, adj, labels
 
 
-for i in range(epochs):
-    tl, ta = ut.train(model,optimizer,features, adj, idx_train,labels)
-    train_loss += [tl]
-    train_acc += [ta]
+# 난수 seed 값, 개수들. 단순 랜덤 int 값 배열 반환이니까~(image id로 사용할)
+def splitDataset(seed, n_train, n_val, n_features):
+    # dataset train/test/val로 나눔
+    np.random.seed(seed)
+    n_train = n_train
+    n_val = n_val
+    n_test = n_features - n_train - n_val  # 500
+    idxs = np.random.permutation(n_features)
+    idx_train = torch.LongTensor(idxs[:n_train])
+    idx_val = torch.LongTensor(idxs[n_train:n_train + n_val])
+    idx_test = torch.LongTensor(idxs[n_train + n_val:])
 
-    if ((i + 1) % print_steps) == 0 or i == 0:
-        tl, ta = ut.evaluate(idx_train, model,features, adj,labels)
-        vl, va = ut.evaluate(idx_train, model,features, adj,labels)
-        val_loss += [vl]
-        val_acc += [va]
+    return idx_train, idx_val, idx_test
 
-        print(
-            'Epochs: {}, Train Loss: {:.3f}, Train Acc: {:.3f}, Validation Loss: {:.3f}, Validation Acc: {:.3f}'.format(
-                i, tl, ta, vl, va))
 
-output = model(features, adj)
+# cuda.. gpu로 보냄
+def toDevice(adj, features, labels, idx_train, idx_val, idx_test):
+    adj = adj.to(device)
+    features = features.to(device)
+    labels = labels.to(device)
+    idx_train = idx_train.to(device)
+    idx_val = idx_val.to(device)
+    idx_test = idx_test.to(device)
 
-# test
-samples = 10
-# torch.randperm : 데이터 셔플
-# idx_sample = idx_test[torch.randperm(len(idx_test))[:samples]]
-idx_sample = idx_test[torch.randperm(len(idx_test))[:samples]]
-print(idx_sample)
 
-idx2lbl = ['0번 그림', '1번 그림', '2번 그림', '3번 그림', '4번 그림', '5번 그림', '6번 그림', '7번 그림'
-    , '8번 그림', '9번 그림', '10번 그림', '11번 그림', '12번 그림', '13번 그림', '14번 그림']
-# 실제 라벨 10개, predicate 결과 10개
+# 학습
+# def step():
+def train(model, optimizer, features, adj, idx_train, labels):
+    t = time.time()
+    model.train()  # model 학습모드로
+    optimizer.zero_grad()
+    output = model(features, adj)  # model에 값 넣음  tensor(1000,15)
+    loss = F.nll_loss(output[idx_train], labels[idx_train])  # loss 함수   tensor, torch.float32
+    acc = accuracy(output[idx_train], labels[idx_train])  # accuracy 파악
+    loss.backward()
+    optimizer.step()
 
-# numpy.argmax() : 주어진 배열에서 가장 높은 값을 가진 인덱스 반환
-# pd.tolist() : 동일 위치(레벨)에 있는 데이터끼리 묶어줌
-df = pd.DataFrame({'Real': [idx2lbl[e] for e in labels[idx_sample].tolist()],
-                   'Pred': [idx2lbl[e] for e in output[idx_sample].argmax(1).tolist()]})
-# [idx2lbl[e] for e in output[idx_sample].argmax(1).tolist()] : 예측한 값이 label의 어디에 속하는지
-# Pred
-# idx_sample 개수만큼의 output을 가져오는데, argmax로 output 값 15개(각 cluster에 대한 값) 중 가장 큰 값을 뽑고, 그 위치를 받음
-# tolist()로 각 feature 별 model의 output 값을 받음
-# for 문을 통해 idx2Lble에서 해당하는 label을 반환받음.
-print(df)
+    return loss.item(), acc
+
+
+# 평가
+def evaluate(idx, model, features, adj, labels):
+    model.eval()
+    output = model(features, adj)  # 모델 돌림
+    loss = F.nll_loss(output[idx], labels[idx])  # 모델이 분류한 값과 label 비교해서 loss 파악
+    acc = accuracy(output[idx], labels[idx])
+
+    return loss.item(), acc
+
+
+# 정확도
+def accuracy(output, labels):  # output : tensor(200, 15), labels : tensor(200,)
+    preds = output.max(1)[1].type_as(labels)  # 비슷하다고 뽑은 것들중에 제일 비슷한 거
+    correct = preds.eq(labels).double()  # tensor (300,)
+    correct = correct.sum()
+    return correct / len(labels)
