--- NodeClassification/NodeClassification/data/cluster.txt
@@ -1,1 +1,0 @@
-13,13,7,4,4,1,9,9,4,9,7,9,9,9,9,13,9,6,9,7,7,13,11,13,2,2,4,4,5,10,1,1,10,4,7,1,0,11,4,0,1,9,1,10,13,12,13,14,2,9,5,11,3,3,13,14,13,6,14,10,14,3,1,2,2,9,1,6,8,7,9,6,7,7,5,10,9,7,4,9,1,1,7,1,4,6,5,4,4,4,4,0,7,0,0,0,0,0,1,9,13,14,14,12,12,2,9,14,13,7,1,7,2,6,6,3,4,4,2,4,10,1,1,1,2,2,13,13,2,11,5,2,5,8,2,13,14,14,2,13,6,11,6,10,12,7,2,13,13,10,1,13,6,1,1,6,11,11,14,5,13,0,11,2,2,14,6,4,0,7,6,2,6,4,14,13,6,1,6,7,2,1,9,4,4,0,1,11,8,3,2,14,7,2,0,2,11,2,0,0,14,3,12,11,10,2,4,10,1,2,11,4,4,6,13,4,14,14,0,5,2,13,1,14,13,11,14,4,13,1,14,4,1,11,6,14,7,0,7,13,0,2,12,4,2,2,4,13,4,14,13,2,12,4,2,2,14,7,4,6,3,13,3,2,11,13,11,4,14,14,13,5,14,11,2,14,6,7,12,6,11,3,14,14,9,0,2,13,4,12,11,2,1,14,4,8,11,11,4,6,14,12,14,2,2,13,13,13,13,14,13,14,14,2,12,2,14,14,10,13,14,13,14,13,13,6,13,9,9,9,7,7,1,1,7,9,13,6,13,13,14,6,13,6,13,13,7,4,9,6,10,13,2,6,13,12,12,9,9,9,2,13,3,14,2,14,12,14,9,7,9,9,9,9,6,10,9,10,1,9,9,9,9,7,7,6,6,9,7,9,13,9,6,14,6,6,2,2,6,14,13,14,9,7,9,9,9,7,7,6,7,6,9,6,0,0,1,6,6,6,14,14,14,6,5,14,13,14,14,13,2,6,14,10,13,6,12,14,14,6,6,6,13,2,3,6,6,6,2,14,14,10,2,6,3,6,13,2,12,3,14,13,2,14,6,8,2,2,13,2,2,2,6,4,4,14,13,13,14,11,6,2,6,6,11,3,2,6,11,1,4,6,9,4,6,4,10,4,4,1,4,4,1,0,1,4,4,1,7,4,4,4,1,10,4,6,4,13,4,4,4,4,1,1,1,7,4,4,4,3,7,4,4,1,4,10,4,9,11,6,4,4,1,4,4,6,9,1,4,4,4,6,0,4,0,0,4,6,1,7,8,1,1,1,1,1,9,1,4,0,4,4,0,9,7,5,3,9,12,6,14,6,14,8,14,13,13,13,13,13,6,11,14,6,13,13,13,2,14,13,12,13,13,3,13,12,12,2,2,6,14,12,8,11,6,2,6,13,13,14,3,6,2,14,14,14,2,2,14,3,6,13,13,6,13,12,6,2,6,4,7,14,14,12,6,14,3,12,10,6,10,6,11,6,2,11,5,2,11,11,2,1,11,0,0,2,6,13,5,2,2,3,5,2,6,4,12,11,2,6,5,5,8,5,8,11,8,8,11,11,6,6,8,8,5,8,5,13,2,12,13,2,2,11,5,5,5,5,5,5,2,12,12,12,10,6,2,3,5,3,14,13,13,7,14,12,13,13,3,14,2,6,3,3,13,6,12,5,5,5,6,2,2,14,11,3,2,2,13,5,0,3,14,13,13,11,11,13,14,13,2,10,3,2,13,3,13,3,3,14,6,6,14,13,2,3,3,3,3,2,2,12,14,14,14,13,14,3,11,3,13,14,2,6,3,6,6,7,2,13,14,12,2,3,5,14,13,5,2,5,5,13,2,5,5,3,2,6,12,14,14,14,14,3,5,11,12,10,12,12,12,13,14,14,10,14,13,12,12,13,11,10,8,11,11,11,6,14,11,14,14,0,5,6,0,0,6,3,11,3,6,13,14,13,13,11,13,14,12,13,13,13,3,2,13,14,3,14,2,2,6,4,14,7,9,5,3,6,11,5,11,14,6,5,10,11,11,6,7,11,3,2,2,13,1,4,6,6,6,0,2,9,12,9,5,9,1,4,3,6,14,6,11,3,11,7,1,14,11,2,6,11,3,4,5,1,6,5,5,5,3,11,2,13,13,11,6,13,3,5,11,11,11,9,9,7,9,9,11,12,6,6,6,6,5,6,10,10,2,3,11,11,2,6,13,3,13,3,10,11,13,13,13,10,10,8,6,2,14,2,14,14,14,2,6,2,6,6,2,6
