--- NodeClassification/NodeClassification/pyvenv.cfg
@@ -1,8 +1,0 @@
-home = C:\Users\Haeun\AppData\Local\Programs\Python\Python39
-implementation = CPython
-version_info = 3.9.7.final.0
-virtualenv = 20.4.7
-include-system-site-packages = false
-base-prefix = C:\Users\Haeun\AppData\Local\Programs\Python\Python39
-base-exec-prefix = C:\Users\Haeun\AppData\Local\Programs\Python\Python39
-base-executable = C:\Users\Haeun\AppData\Local\Programs\Python\Python39\python.exe
